Various changes involving unicode precedence lists.
Make an object to encapsulate the dynarr used internally to represent
unicode precedence lists, and change functions to use it.  Remove the
code that implements a single language-specific precedence list, in
preparation for making a buffer-local precedence list.  Allow tags as
well as charsets to be present in precedence lists, calling a Lisp
function `charset-tag-to-charset-list' to normalize tags to lists of
charsets.

Additions to charset tags:
-- list tags, with :list property, which expand to an in-order list
   of charsets/tags, rather than matching charsets with the tag
   given
-- combined tags: instead of a tag, a list of tags can be given, and
   the result is the intersection of the results of the individual
   tags, essentially 'this and that'

New function `charset-or-charset-name-p' for indicating whether its
argument is something that would be acceptable to functions that
take a charset.

Define list tags for Latin, Chinese, and Japanese, and use them
to construct a more sensible default for `set-default-unicode-precedence-list'.

Functions `charset-tag-p' and `charset-or-charset-tag-p'.

Rebuild dependencies.